added docs, updated tests, psr-4, updated dependencies, changed namespaces (Minetro=>Contributte), php 7.1+, QA + CS
