+++ .gitattributes
@@ -1,6 +1,10 @@
-.gitattributes export-ignore
-.gitignore export-ignore
-.travis.yml export-ignore
-composer.lock export-ignore
-tests/ export-ignore
-*.md export-ignore
+/.gitattributes export-ignore
+/.gitignore export-ignore
+/.travis.yml export-ignore
+/phpstan.neon export-ignore
+/tests/ export-ignore
+/.editorconfig export-ignore
+/composer.lock export-ignore
+/README.md export-ignore
+/LICENSE export-ignore
+/.docs export-ignore
