+++ composer.json
@@ -1,23 +1,54 @@
 {
-  "name": "minetro/events",
+  "name": "contributte/events",
   "description": "Simple events for Nette",
   "type": "library",
-  "license": ["BSD-3-Clause"],
-  "homepage": "https://github.com/minetro/events",
+  "license": ["MIT"],
+  "homepage": "https://github.com/contributte/events",
   "authors": [
     {
       "name": "Milan Felix Sulc",
-      "homepage": "http://jfx.cz"
+      "homepage": "https://f3l1x.io"
     }
   ],
   "require": {
-    "nette/di": ">=2.3.0,<2.4.0"
+    "php": ">=7.1",
+    "nette/di": "~2.4.12"
   },
   "require-dev": {
-    "nette/tester": "~1.4.0",
-    "nette/bootstrap": ">=2.3.0,<2.4.0"
+    "ninjify/qa": "^0.8.0",
+    "ninjify/nunjuck": "^0.2.0",
+    "nette/bootstrap": "~2.4.6"
   },
   "autoload": {
-    "classmap": ["src/"]
+    "psr-4": {
+      "Contributte\\Events\\": "src/"
+    }
+  },
+  "autoload-dev": {
+    "psr-4": {
+      "Tests\\Fixtures\\": "tests/fixtures/"
+    }
+  },
+  "minimum-stability": "dev",
+  "prefer-stable": true,
+  "scripts": {
+    "qa": [
+      "linter src tests",
+      "codesniffer src tests"
+    ],
+    "tester": [
+      "tester -s -p php --colors 1 -C tests/cases"
+    ],
+    "coverage": [
+      "tester -s -p php --colors 1 -C -d extension=xdebug.so --coverage ./coverage.xml --coverage-src ./src tests/Events"
+    ],
+    "phpstan-install": [
+      "mkdir -p temp/phpstan",
+      "composer require -d temp/phpstan phpstan/phpstan:0.9.2",
+      "composer require -d temp/phpstan phpstan/phpstan-nette:0.9"
+    ],
+    "phpstan": [
+      "temp/phpstan/vendor/bin/phpstan analyse -l max -c phpstan.neon src"
+    ]
   }
 }
