+++ ruleset.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0"?>
+<ruleset name="Contributte">
+    <!-- Extending rulesets -->
+    <rule ref="vendor/ninjify/coding-standard/contributte.xml"/>
+
+    <!-- My rules -->
+    <rule ref="SlevomatCodingStandard.Files.TypeNameMatchesFileName">
+        <properties>
+            <property name="rootNamespaces" type="array" value="
+                src=>Contributte\Events,
+                tests/fixtures=>Tests\Fixtures
+            "/>
+        </properties>
+    </rule>
+
+    <!-- Exclude folders -->
+    <exclude-pattern>/tests/tmp</exclude-pattern>
+</ruleset>
