+++ tests/.gitignore
@@ -1,0 +1,10 @@
+# Folders - recursive
+*.expected
+*.actual
+
+# Folders
+/tmp
+
+# Files
+/*.log
+/*.html
