+++ tests/fixtures/Mocks/DI/MockContainer.php
@@ -1,0 +1,10 @@
+<?php declare(strict_types = 1);
+
+namespace Tests\Fixtures\Mocks\DI;
+
+use Nette\DI\Container;
+
+final class MockContainer extends Container
+{
+
+}
