+++ tests/fixtures/Mocks/Subscribers/MockSubscriber.php
@@ -1,0 +1,30 @@
+<?php declare(strict_types = 1);
+
+namespace Tests\Fixtures\Mocks\Subscribers;
+
+use Contributte\Events\EventsManager;
+use Contributte\Events\EventsSubscriber;
+
+final class MockSubscriber implements EventsSubscriber
+{
+
+	/** @var int */
+	public $calls = 0;
+
+	/** @var int */
+	public $innerCalls = 0;
+
+	public function onEvents(EventsManager $em): void
+	{
+		$this->calls++;
+		$em->on('tester.test1', function () {
+			$this->innerCalls++;
+			return ['tester.test1' => func_get_args()];
+		});
+		$em->on('tester.test2', function () {
+			$this->innerCalls++;
+			return ['tester.test2' => func_get_args()];
+		});
+	}
+
+}
