--- tests/php-win.ini
@@ -1,6 +1,0 @@
-[PHP]
-extension_dir=./ext
-extension=php_openssl.dll
-
-[XDebug]
-zend_extension = "php_xdebug.dll"
