--- tests/tester
@@ -1,3 +1,0 @@
-#!/usr/bin/env sh
-sh ../vendor/bin/tester Events -s -j 40 -log events.log $@
-rm "./tmp" -f -R
